That was fun. Ok things are split apart and it builds again. Needs a few hours more cleanup but its a start.
