+++ components/otaupdate/CMakeLists.txt
@@ -1,0 +1,3 @@
+idf_component_register(SRCS "ota_utils.c"
+                    INCLUDE_DIRS .)
+project(otaupdate)
