+++ components/otaupdate/update_public_key.pem
