+++ components/stsdk/caps_carbonMonoxideDetector.c
