+++ components/stsdk/caps_carbonMonoxideDetector.h
