+++ components/stsdk/caps_contactSensor.c
