+++ components/stsdk/caps_contactSensor.h
