+++ components/stsdk/caps_momentary.c
