+++ components/stsdk/caps_momentary.h
