+++ components/stsdk/caps_refresh.c
