+++ components/stsdk/caps_refresh.h
