+++ components/stsdk/caps_securitySystem.c
