+++ components/stsdk/caps_smokeDetector.c
