+++ components/stsdk/caps_smokeDetector.h
