+++ components/stsdk/caps_switch.c
