+++ components/stsdk/caps_switch.h
