+++ components/stsdk/device_info.json
