+++ components/stsdk/onboarding_config.json
