+++ components/stsdk/stsdk_main.h
@@ -1,10 +1,9 @@
 /**
- *  @file    alarmdecoder_stsdk_app_esp32.h
+ *  @file    stsdk_main.h
  *  @author  Sean Mathews <coder@f34r.com>
- *  @date    02/20/2020
- *  @version 1.0
+ *  @date    09/18/2020
  *
- *  @brief AlarmDecoder IoT embedded network appliance for SmartThings
+ *  @brief SmartThings Direct Attached Device using STSDK
  *
  *  @copyright Copyright (C) 2020 Nu Tech Software Solutions, Inc.
  *
@@ -21,27 +20,58 @@
  *  limitations under the License.
  *
  */
-#ifndef AlarmDecoder_app_h
-#define AlarmDecoder_app_h
+
+#ifndef stsdk_main_h
+#define stsdk_main_h
+
+
+#include "st_dev.h"
+#include "device_control.h"
+
+// device types
+#include "caps_refresh.h"
+#include "caps_switch.h"
+#include "caps_securitySystem.h"
+#include "caps_momentary.h"
+#include "caps_smokeDetector.h"
+#include "caps_tamperAlert.h"
+#include "caps_contactSensor.h"
+#include "caps_carbonMonoxideDetector.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-void alarmdecoder_start(iot_status_t *, IOT_CTX**);
+void stsdk_init(void);
 void button_event(IOT_CAP_HANDLE *handle, int type, int count);
-static void iot_noti_cb(iot_noti_data_t *noti_data, void *noti_usr_data);
-static void connection_start(void);
-static void capability_init();
-static void ota_polling_task_func(void *arg);
+void iot_noti_cb(iot_noti_data_t *noti_data, void *noti_usr_data);
+void connection_start(void);
+void capability_init();
 void cap_current_version_init_cb(IOT_CAP_HANDLE *handle, void *usr_data);
 void update_firmware_cmd_cb(IOT_CAP_HANDLE *handle, iot_cap_cmd_data_t *cmd_data, void *usr_data);
 void cap_health_check_init_cb(IOT_CAP_HANDLE *handle, void *usr_data);
 void ping_cmd_cb(IOT_CAP_HANDLE *handle, iot_cap_cmd_data_t *cmd_data, void *usr_data);
 void refresh_cmd_cb(IOT_CAP_HANDLE *handle, iot_cap_cmd_data_t *cmd_data, void *usr_data);
-void send_to_ad2(char *buf);
+
+
+extern iot_status_t g_iot_status;
+extern iot_stat_lv_t g_iot_stat_lv;
+
+extern caps_switch_data_t *cap_switch_data;
+extern caps_contactSensor_data_t *cap_contactSensor_data_chime;
+extern caps_momentary_data_t *cap_momentary_data_chime;
+
+extern caps_securitySystem_data_t *cap_securitySystem_data;
+
+extern caps_smokeDetector_data_t *cap_smokeDetector_data;
+extern caps_momentary_data_t *cap_momentary_data_fire;
+
+extern IOT_CAP_HANDLE *ota_cap_handle;
+extern IOT_CAP_HANDLE *healthCheck_cap_handle;
+extern IOT_CAP_HANDLE *refresh_cap_handle;
+
+extern int noti_led_mode;
+
 #ifdef __cplusplus
 }
 #endif
-
-#endif
+#endif // stsdk_main_h
