+++ components/twilio/twilio.h
@@ -2,7 +2,6 @@
  *  @file    twilio.h
  *  @author  Sean Mathews <coder@f34r.com>
  *  @date    09/12/2020
- *  @version 1.0
  *
  *  @brief Simple commands to post to api.twilio.com
  *
