+++ main/CMakeLists.txt
@@ -1,21 +1,12 @@
-idf_component_register(SRCS "alarmdecoder_stsdk_app_esp32.cpp"
+idf_component_register(SRCS "alarmdecoder_main.cpp"
                             "device_control.c"
                             "iot_cli_cmd.c"
                             "iot_uart_cli.c"
                             "ota_util.c"
-                            "caps_switch.c"
-                            "caps_tamperAlert.c"
-                            "caps_smokeDetector.c"
-                            "caps_securitySystem.c"
-                            "caps_momentary.c"
-                            "caps_contactSensor.c"
-                            "caps_carbonMonoxideDetector.c"
                             "ad2_utils.c"
-                            "caps_refresh.c"
-                    EMBED_FILES "device_info.json"
-                                "onboarding_config.json"
                     REQUIRES idf::alarmdecoder-api
                     REQUIRES idf::twilio
+		    REQUIRES idf::stsdk
                     )
 
 set(STDK_IOT_CORE_USE_DEFINED_CONFIG "y")
