+++ main/alarmdecoder_main.cpp
@@ -1,10 +1,10 @@
  /**
- *  @file    alarmdecoder_stsdk_app_esp32.cpp
+ *  @file    alarmdecoder_main.cpp
  *  @author  Sean Mathews <coder@f34r.com>
  *  @date    02/20/2020
  *  @version 1.0
  *
- *  @brief AlarmDecoder IoT embedded network appliance for SmartThings
+ *  @brief AlarmDecoder IoT embedded network appliance
  *
  *  @copyright Copyright (C) 2020 Nu Tech Software Solutions, Inc.
  *
@@ -33,27 +33,6 @@
 extern "C" {
 #endif
 
-// STSDK include
-#include "st_dev.h"
-#include "device_control.h"
-
-// device types
-#include "caps_refresh.h"
-#include "caps_switch.h"
-#include "caps_securitySystem.h"
-#include "caps_momentary.h"
-#include "caps_smokeDetector.h"
-#include "caps_tamperAlert.h"
-#include "caps_contactSensor.h"
-#include "caps_carbonMonoxideDetector.h"
-
-// OTA updates
-#include "ota_util.h"
-
-// ST CLI
-#include "iot_uart_cli.h"
-#include "iot_cli_cmd.h"
-
 // esp includes
 #include "driver/uart.h"
 #include <lwip/netdb.h>
@@ -61,14 +40,26 @@ extern "C" {
 #include "nvs_flash.h"
 #include "nvs.h"
 
+// OTA updates
+#include "ota_util.h"
+
+// CLI interface
+#include "iot_uart_cli.h"
+#include "iot_cli_cmd.h"
+
 // common settings
 #include "ad2_settings.h"
 
 // AD2IoT include
-#include "alarmdecoder_stsdk_app_esp32.h"
+#include "alarmdecoder_main.h"
 
 // common utils
 #include "ad2_utils.h"
+
+// SmartThings direct attached device support
+#ifdef CONFIG_STDK_IOT_CORE
+#include "stsdk_main.h"
+#endif
 
 // twilio support
 #if CONFIG_TWILIO_CLIENT
@@ -80,42 +71,15 @@ extern "C" {
  */
 static const char *TAG = "AD2_IoT";
 
+// global AlarmDecoder parser class instance
 static AlarmDecoderParser AD2Parse;
 
 // AD2 device connection settings
 int g_ad2_client_handle = -1;
 uint8_t g_ad2_mode = 0;
 
-// onboarding_config_start is null-terminated string
-extern const uint8_t onboarding_config_start[]  asm("_binary_onboarding_config_json_start");
-extern const uint8_t onboarding_config_end[]    asm("_binary_onboarding_config_json_end");
-
-// device_info_start is null-terminated string
-extern const uint8_t device_info_start[]        asm("_binary_device_info_json_start");
-extern const uint8_t device_info_end[]          asm("_binary_device_info_json_end");
-
-static iot_status_t g_iot_status = IOT_STATUS_IDLE;
-static iot_stat_lv_t g_iot_stat_lv;
-
-IOT_CTX* ctx = NULL;
-
-//#define SET_PIN_NUMBER_CONFRIM
-
-static int noti_led_mode = LED_ANIMATION_MODE_IDLE;
-
-static caps_switch_data_t *cap_switch_data;
-static caps_contactSensor_data_t *cap_contactSensor_data_chime;
-static caps_momentary_data_t *cap_momentary_data_chime;
-
-static caps_securitySystem_data_t *cap_securitySystem_data;
-
-static caps_smokeDetector_data_t *cap_smokeDetector_data;
-static caps_momentary_data_t *cap_momentary_data_fire;
-
-TaskHandle_t ota_task_handle = NULL;
-IOT_CAP_HANDLE *ota_cap_handle = NULL;
-IOT_CAP_HANDLE *healthCheck_cap_handle = NULL;
-IOT_CAP_HANDLE *refresh_cap_handle = NULL;
+// Device LED mode
+int noti_led_mode = LED_ANIMATION_MODE_IDLE;
 
 /**
  * AlarmDecoder callbacks.
@@ -158,12 +122,14 @@ void my_ON_READY_CHANGE_CB(std::string *msg, AD2VirtualPartitionState *s) {
 void my_ON_ARM_CB(std::string *msg, AD2VirtualPartitionState *s) {
   ESP_LOGI(TAG, "ON_ARM: READY(%i) EXIT(%i) HOME(%i) AWAY(%i)", s->ready, s->exit_now, s->armed_home, s->armed_away);
 
+#ifdef CONFIG_STDK_IOT_CORE
   if (s->armed_home)
     cap_securitySystem_data->set_securitySystemStatus_value(cap_securitySystem_data, caps_helper_securitySystem.attr_securitySystemStatus.value_armedStay);
   else
     cap_securitySystem_data->set_securitySystemStatus_value(cap_securitySystem_data, caps_helper_securitySystem.attr_securitySystemStatus.value_armedAway);
 
   cap_securitySystem_data->attr_securitySystemStatus_send(cap_securitySystem_data);
+#endif
 }
 
 /**
@@ -173,9 +139,11 @@ void my_ON_ARM_CB(std::string *msg, AD2VirtualPartitionState *s) {
 void my_ON_DISARM_CB(std::string *msg, AD2VirtualPartitionState *s) {
   ESP_LOGI(TAG, "ON_DISARM: READY(%i)", s->ready);
 
+#ifdef CONFIG_STDK_IOT_CORE
   const char *new_value = caps_helper_securitySystem.attr_securitySystemStatus.value_disarmed;
   cap_securitySystem_data->set_securitySystemStatus_value(cap_securitySystem_data, new_value);
   cap_securitySystem_data->attr_securitySystemStatus_send(cap_securitySystem_data);
+#endif
 }
 
 /**
@@ -186,12 +154,14 @@ void my_ON_DISARM_CB(std::string *msg, AD2VirtualPartitionState *s) {
  */
 void my_ON_CHIME_CHANGE_CB(std::string *msg, AD2VirtualPartitionState *s) {
   ESP_LOGI(TAG, "ON_CHIME_CHANGE: CHIME(%i)", s->chime_on);
+#ifdef CONFIG_STDK_IOT_CORE
   if ( s->chime_on)
     cap_contactSensor_data_chime->set_contact_value(cap_contactSensor_data_chime, caps_helper_contactSensor.attr_contact.value_open);
   else
     cap_contactSensor_data_chime->set_contact_value(cap_contactSensor_data_chime, caps_helper_contactSensor.attr_contact.value_closed);
 
   cap_contactSensor_data_chime->attr_contact_send(cap_contactSensor_data_chime);
+#endif
 
 // FIXME: Testing.. todo: cleanup compress simplify
 #if CONFIG_TWILIO_CLIENT
@@ -238,6 +208,7 @@ void my_ON_CHIME_CHANGE_CB(std::string *msg, AD2VirtualPartitionState *s) {
  */
 void my_ON_FIRE_CB(std::string *msg, AD2VirtualPartitionState *s) {
   ESP_LOGI(TAG, "ON_FIRE_CB: FIRE(%i)", s->fire_alarm);
+#ifdef CONFIG_STDK_IOT_CORE
   if ( s->fire_alarm ) {
     cap_smokeDetector_data->set_smoke_value(cap_smokeDetector_data, caps_helper_smokeDetector.attr_smoke.value_detected);
   } else {
@@ -245,6 +216,7 @@ void my_ON_FIRE_CB(std::string *msg, AD2VirtualPartitionState *s) {
   }
 
   cap_smokeDetector_data->attr_smoke_send(cap_smokeDetector_data);
+#endif
 }
 
 /**
@@ -657,25 +629,7 @@ void app_main()
     AD2Parse.setCB_ON_FIRE(my_ON_FIRE_CB);
 
     //// SmartThings setup
-    unsigned char *onboarding_config = (unsigned char *) onboarding_config_start;
-    unsigned int onboarding_config_len = onboarding_config_end - onboarding_config_start;
-    unsigned char *device_info = (unsigned char *) device_info_start;
-    unsigned int device_info_len = device_info_end - device_info_start;
-
-    int iot_err;
-
-    // create a iot context
-    ctx = st_conn_init(onboarding_config, onboarding_config_len, device_info, device_info_len);
-    if (ctx != NULL) {
-        iot_err = st_conn_set_noti_cb(ctx, iot_noti_cb, NULL);
-        if (iot_err)
-            ESP_LOGE(TAG, "fail to set notification callback function");
-    } else {
-        ESP_LOGE(TAG, "fail to create the iot_context");
-    }
-
-    // Add AD2 capabilies
-    capability_init();
+    stsdk_init();
 
     // Init onboarding button and LED
     iot_gpio_init();
@@ -697,347 +651,6 @@ void app_main()
 
 
 
-
-
-
-/*
- * Smartthings
- *
- */
-
-static int get_switch_state(void)
-{
-    const char* switch_value = cap_switch_data->get_switch_value(cap_switch_data);
-    int switch_state = SWITCH_OFF;
-
-    if (!switch_value) {
-        return -1;
-    }
-
-    if (!strcmp(switch_value, caps_helper_switch.attr_switch.value_on)) {
-        switch_state = SWITCH_ON;
-    } else if (!strcmp(switch_value, caps_helper_switch.attr_switch.value_off)) {
-        switch_state = SWITCH_OFF;
-    }
-    return switch_state;
-}
-
-/**
- * callback from cloud to arm away. We modify to report back to the
- * cloud our current state not an armed away state. If/when arm away state
- * changes it will be sent.
- */
-static void cap_securitySystem_arm_away_cmd_cb(struct caps_securitySystem_data *caps_data)
-{
-    const char *init_value = caps_helper_securitySystem.attr_securitySystemStatus.value_disarmed;
-    cap_securitySystem_data->set_securitySystemStatus_value(cap_securitySystem_data, init_value);
-
-    // TODO: uint32_t amask = 0xffffff7f;
-    // TODO: AD2VirtualPartitionState * s = AD2Parse.getAD2PState(&amask);
-
-    // send back our current state
-    //cap_securitySystem_data->attr_securitySystemStatus_send(cap_securitySystem_data);
-
-    // Get user code
-    char code[7];
-    ad2_get_nv_slot_key_string(CODES_CONFIG_KEY, AD2_DEFAULT_CODE_SLOT, code, sizeof(code));
-
-    // Get the address/partition mask
-    // Message format KXXYYYYZ
-    char msg[9] = {0};
-    int address = -1;
-    ad2_get_nv_slot_key_int(VPADDR_CONFIG_KEY, AD2_DEFAULT_VPA_SLOT, &address);
-    snprintf(msg, sizeof(msg), "K%02i%s%s", address, code, "3");
-    ESP_LOGI(TAG ,"Sending ARM AWAY command");
-    send_to_ad2(msg);
-}
-
-static void cap_securitySystem_arm_stay_cmd_cb(struct caps_securitySystem_data *caps_data)
-{
-    const char *init_value = caps_helper_securitySystem.attr_securitySystemStatus.value_disarmed;
-    cap_securitySystem_data->set_securitySystemStatus_value(cap_securitySystem_data, init_value);
-
-    // TODO: uint32_t amask = 0xffffff7f;
-    // TODO: AD2VirtualPartitionState * s = AD2Parse.getAD2PState(&amask);
-
-    // send back our current state
-    //cap_securitySystem_data->attr_securitySystemStatus_send(cap_securitySystem_data);
-
-    // Get user code
-    char code[7];
-    ad2_get_nv_slot_key_string(CODES_CONFIG_KEY, AD2_DEFAULT_CODE_SLOT, code, sizeof(code));
-
-    // Get the address/partition mask
-    // Message format KXXYYYYZ
-    char msg[9] = {0};
-    int address = -1;
-    ad2_get_nv_slot_key_int(VPADDR_CONFIG_KEY, AD2_DEFAULT_VPA_SLOT, &address);
-    snprintf(msg, sizeof(msg), "K%02i%s%s", address, code, "3");
-    ESP_LOGI(TAG, "Sending ARM STAY command");
-    send_to_ad2(msg);
-}
-
-static void cap_securitySystem_disarm_cmd_cb(struct caps_securitySystem_data *caps_data)
-{
-    const char *init_value = caps_helper_securitySystem.attr_securitySystemStatus.value_disarmed;
-    cap_securitySystem_data->set_securitySystemStatus_value(cap_securitySystem_data, init_value);
-
-    // TODO: uint32_t amask = 0xffffff7f;
-    // TODO: AD2VirtualPartitionState * s = AD2Parse.getAD2PState(&amask);
-
-    // send back our current state
-    //cap_securitySystem_data->attr_securitySystemStatus_send(cap_securitySystem_data);
-
-    // Get user code
-    char code[7] = {0};
-    ad2_get_nv_slot_key_string(CODES_CONFIG_KEY, AD2_DEFAULT_CODE_SLOT, code, sizeof(code));
-
-    // Get the address/partition mask
-    // Message format KXXYYYYZ
-    char msg[9] = {0};
-    int address = -1;
-    ad2_get_nv_slot_key_int(VPADDR_CONFIG_KEY, AD2_DEFAULT_VPA_SLOT, &address);
-    snprintf(msg, sizeof(msg), "K%02i%s%s", address, code, "1");
-    ESP_LOGI(TAG, "Sending DISARM command");
-    send_to_ad2(msg);
-}
-
-static void cap_securitySystem_chime_cmd_cb(struct caps_momentary_data *caps_data)
-{
-    // TODO: uint32_t amask = 0xffffff7f;
-    // TODO: AD2VirtualPartitionState * s = AD2Parse.getAD2PState(&amask);
-
-    // Get user code
-    char code[7];
-    ad2_get_nv_slot_key_string(CODES_CONFIG_KEY, AD2_DEFAULT_CODE_SLOT, code, sizeof(code));
-
-    // Get the address/partition mask
-    // Message format KXXYYYYZ
-    char msg[9] = {0};
-    int address = -1;
-    ad2_get_nv_slot_key_int(VPADDR_CONFIG_KEY, AD2_DEFAULT_VPA_SLOT, &address);
-    snprintf(msg, sizeof(msg), "K%02i%s%s", address, code, "9");
-    ESP_LOGI(TAG, "Sending CHIME toggle command");
-    send_to_ad2(msg);
-}
-
-static void cap_fire_cmd_cb(struct caps_momentary_data *caps_data)
-{
-    // Get user code
-    char code[7];
-    ad2_get_nv_slot_key_string(CODES_CONFIG_KEY, AD2_DEFAULT_CODE_SLOT, code, sizeof(code));
-
-    // Get the address/partition mask
-    // Message format KXXYYYYZ
-    char msg[9] = {0};
-    int address = -1;
-    ad2_get_nv_slot_key_int(VPADDR_CONFIG_KEY, AD2_DEFAULT_VPA_SLOT, &address);
-    snprintf(msg, sizeof(msg), "K%02i%s", address, "\004\004\004");
-    ESP_LOGI(TAG, "Sending CHIME toggle command");
-#if 0 // DANGER trip FIRE PANIC alarm on panel.
-    send_to_ad2(msg);
-#endif
-}
-
-static void cap_switch_cmd_cb(struct caps_switch_data *caps_data)
-{
-    int switch_state = get_switch_state();
-    change_switch_state(switch_state);
-}
-
-/**
- * Initialize all SmartThings capabilities.
- */
-static void capability_init()
-{
-    ESP_LOGI(TAG, "capability_init");
-    int iot_err;
-
-    // FIXME: dummy switch tied to physical button on ESP32
-    cap_switch_data = caps_switch_initialize(ctx, "trash", NULL, NULL);
-    if (cap_switch_data) {
-        // FIXME: const char *init_value = caps_helper_switch.attr_switch.value_on;
-        cap_switch_data->cmd_on_usr_cb = cap_switch_cmd_cb;
-        cap_switch_data->cmd_off_usr_cb = cap_switch_cmd_cb;
-    }
-
-    // refresh device type init
-    refresh_cap_handle = st_cap_handle_init(ctx, "main", "refresh", NULL, NULL);
-    if (refresh_cap_handle) {
-		iot_err = st_cap_cmd_set_cb(refresh_cap_handle, "refresh", refresh_cmd_cb, NULL);
-		if (iot_err)
-            ESP_LOGE(TAG, "fail to set cmd_cb for refresh");
-    }
-
-    // healthCheck capabilities init
-	healthCheck_cap_handle = st_cap_handle_init(ctx, "main", "healthCheck", cap_health_check_init_cb, NULL);
-    if (healthCheck_cap_handle) {
-		iot_err = st_cap_cmd_set_cb(healthCheck_cap_handle, "ping", refresh_cmd_cb, NULL);
-		if (iot_err)
-            ESP_LOGE(TAG, "fail to set cmd_cb for healthCheck");
-    }
-
-    // firmwareUpdate capabilities init
-	ota_cap_handle = st_cap_handle_init(ctx, "main", "firmwareUpdate", cap_current_version_init_cb, NULL);
-    if (ota_cap_handle) {
-		iot_err = st_cap_cmd_set_cb(ota_cap_handle, "updateFirmware", update_firmware_cmd_cb, NULL);
-		if (iot_err)
-			ESP_LOGE(TAG, "fail to set cmd_cb for updateFirmware");
-    }
-
-    // securitySystem device type init
-    cap_securitySystem_data = caps_securitySystem_initialize(ctx, "main", NULL, NULL);
-    if (cap_securitySystem_data) {
-        const char *init_value = caps_helper_securitySystem.attr_securitySystemStatus.value_disarmed;
-        cap_securitySystem_data->set_securitySystemStatus_value(cap_securitySystem_data, init_value);
-        cap_securitySystem_data->cmd_armAway_usr_cb = cap_securitySystem_arm_away_cmd_cb;
-        cap_securitySystem_data->cmd_armStay_usr_cb = cap_securitySystem_arm_stay_cmd_cb;
-        cap_securitySystem_data->cmd_disarm_usr_cb = cap_securitySystem_disarm_cmd_cb;
-    }
-
-    // Chime contact sensor
-    cap_contactSensor_data_chime = caps_contactSensor_initialize(ctx, "chime", NULL, NULL);
-    if (cap_contactSensor_data_chime) {
-        const char *contact_init_value = caps_helper_contactSensor.attr_contact.value_open;
-        cap_contactSensor_data_chime->set_contact_value(cap_contactSensor_data_chime, contact_init_value);
-    }
-    // Chime Momentary button
-    cap_momentary_data_chime = caps_momentary_initialize(ctx, "chime", NULL, NULL);
-    if (cap_momentary_data_chime) {
-        cap_momentary_data_chime->cmd_push_usr_cb = cap_securitySystem_chime_cmd_cb;
-    }
-
-    // Smoke Detector
-    cap_smokeDetector_data = caps_smokeDetector_initialize(ctx, "fire", NULL, NULL);
-    if (cap_smokeDetector_data) {
-        const char *smoke_init_value = caps_helper_smokeDetector.attr_smoke.value_clear;
-        cap_smokeDetector_data->set_smoke_value(cap_smokeDetector_data, smoke_init_value);
-    }
-
-    // Fire Alarm Momentary button
-    cap_momentary_data_fire = caps_momentary_initialize(ctx, "fire", NULL, NULL);
-    if (cap_momentary_data_fire) {
-        cap_momentary_data_fire->cmd_push_usr_cb = cap_fire_cmd_cb;
-    }
-}
-
-static void iot_status_cb(iot_status_t status,
-                          iot_stat_lv_t stat_lv, void *usr_data)
-{
-    g_iot_status = status;
-    g_iot_stat_lv = stat_lv;
-
-    ESP_LOGI(TAG, "iot_status_cb %d, stat: %d", g_iot_status, g_iot_stat_lv);
-
-    switch(status)
-    {
-        case IOT_STATUS_NEED_INTERACT:
-            noti_led_mode = LED_ANIMATION_MODE_FAST;
-            break;
-        case IOT_STATUS_IDLE:
-        case IOT_STATUS_CONNECTING:
-            noti_led_mode = LED_ANIMATION_MODE_IDLE;
-            // FIXME: send update on all caps to cloud
-            change_switch_state(get_switch_state());
-            break;
-        default:
-            break;
-    }
-}
-
-#if defined(SET_PIN_NUMBER_CONFRIM)
-void* pin_num_memcpy(void *dest, const void *src, unsigned int count)
-{
-    unsigned int i;
-    for (i = 0; i < count; i++)
-        *((char*)dest + i) = *((char*)src + i);
-    return dest;
-}
-#endif
-
-static void connection_start(void)
-{
-    iot_pin_t *pin_num = NULL;
-    int err;
-
-#if defined(SET_PIN_NUMBER_CONFRIM)
-    pin_num = (iot_pin_t *) malloc(sizeof(iot_pin_t));
-    if (!pin_num)
-        ESP_LOGE(TAG, "failed to malloc for iot_pin_t");
-
-    // to decide the pin confirmation number(ex. "12345678"). It will use for easysetup.
-    //    pin confirmation number must be 8 digit number.
-    pin_num_memcpy(pin_num, "12345678", sizeof(iot_pin_t));
-#endif
-
-    // process on-boarding procedure. There is nothing more to do on the app side than call the API.
-    err = st_conn_start(ctx, (st_status_cb)&iot_status_cb, IOT_STATUS_ALL, NULL, pin_num);
-    if (err) {
-        ESP_LOGE(TAG, "fail to start connection. err:%d", err);
-    }
-    if (pin_num) {
-        free(pin_num);
-    }
-}
-
-static void connection_start_task(void *arg)
-{
-    connection_start();
-    vTaskDelete(NULL);
-}
-
-static void iot_noti_cb(iot_noti_data_t *noti_data, void *noti_usr_data)
-{
-    ESP_LOGI(TAG, "Notification message received");
-
-    if (noti_data->type == IOT_NOTI_TYPE_DEV_DELETED) {
-        ESP_LOGI(TAG, "[device deleted]");
-    } else if (noti_data->type == IOT_NOTI_TYPE_RATE_LIMIT) {
-        ESP_LOGI(TAG, "[rate limit] Remaining time:%d, sequence number:%d",
-               noti_data->raw.rate_limit.remainingTime, noti_data->raw.rate_limit.sequenceNumber);
-    }
-}
-
-void button_event(IOT_CAP_HANDLE *handle, int type, int count)
-{
-    if (type == BUTTON_SHORT_PRESS) {
-        ESP_LOGI(TAG, "Button short press, count: %d", count);
-        switch(count) {
-            case 1:
-                if (g_iot_status == IOT_STATUS_NEED_INTERACT) {
-                    st_conn_ownership_confirm(ctx, true);
-                    noti_led_mode = LED_ANIMATION_MODE_IDLE;
-                    change_switch_state(get_switch_state());
-                } else {
-                    if (get_switch_state() == SWITCH_ON) {
-                        change_switch_state(SWITCH_OFF);
-                        cap_switch_data->set_switch_value(cap_switch_data, caps_helper_switch.attr_switch.value_off);
-                        cap_switch_data->attr_switch_send(cap_switch_data);
-                    } else {
-                        change_switch_state(SWITCH_ON);
-                        cap_switch_data->set_switch_value(cap_switch_data, caps_helper_switch.attr_switch.value_on);
-                        cap_switch_data->attr_switch_send(cap_switch_data);
-                    }
-                }
-                break;
-            case 5:
-                /* clean-up provisioning & registered data with reboot option*/
-                st_conn_cleanup(ctx, true);
-
-                break;
-            default:
-                led_blink(get_switch_state(), 100, count);
-                break;
-        }
-    } else if (type == BUTTON_LONG_PRESS) {
-        ESP_LOGI(TAG, "Button long press, iot_status: %d", g_iot_status);
-        led_blink(get_switch_state(), 100, 3);
-        st_conn_cleanup(ctx, false);
-        xTaskCreate(connection_start_task, "connection_task", 2048, NULL, tskIDLE_PRIORITY+2, NULL);
-    }
-}
-
-
 /*
  * OTA update support
  */
@@ -1055,7 +668,7 @@ void cap_available_version_set(char *available_version)
 	/* Setup switch on state */
 	init_evt = st_cap_attr_create_string("availableVersion", available_version, NULL);
 
-	/* Send switch on event */
+	/* Send avail version to ota cap handler */
 	sequence_no = st_cap_attr_send(ota_cap_handle, evt_num, &init_evt);
 	if (sequence_no < 0)
 		ESP_LOGE(TAG, "fail to send init_data");
@@ -1097,30 +710,6 @@ void cap_current_version_init_cb(IOT_CAP_HANDLE *handle, void *usr_data)
 	st_cap_attr_free(init_evt);
 }
 
-static void _task_fatal_error()
-{
-	ota_task_handle = NULL;
-	ESP_LOGE(TAG, "Exiting task due to fatal error...");
-	(void)vTaskDelete(NULL);
-
-	while (1) {
-		;
-	}
-}
-static void ota_task_func(void * pvParameter)
-{
-	ESP_LOGI(TAG, "Starting OTA");
-
-	esp_err_t ret = ota_https_update_device();
-	if (ret != ESP_OK) {
-		ESP_LOGE(TAG, "Firmware Upgrades Failed (%d)", ret);
-		_task_fatal_error();
-	}
-
-	ESP_LOGI(TAG, "Prepare to restart system!");
-	esp_restart();
-}
-
 void refresh_cmd_cb(IOT_CAP_HANDLE *handle,
 			iot_cap_cmd_data_t *cmd_data, void *usr_data)
 {
@@ -1154,9 +743,7 @@ void refresh_cmd_cb(IOT_CAP_HANDLE *handle,
 void update_firmware_cmd_cb(IOT_CAP_HANDLE *handle,
 			iot_cap_cmd_data_t *cmd_data, void *usr_data)
 {
-	ota_nvs_flash_init();
-
-	xTaskCreate(&ota_task_func, "ota_task_func", 8096, NULL, tskIDLE_PRIORITY+2, &ota_task_handle);
+  do_ota_update();
 }
 
 static void ota_polling_task_func(void *arg)
