+++ main/component.mk
@@ -1,6 +1,10 @@
 #
-# "main" pseudo-component makefile.
+# Component Makefile
 #
-# (Uses default behaviour of compiling all source files in directory, adding 'include' to include path.)
+# This Makefile should, at the very least, just include $(SDK_PATH)/Makefile. By default,
+# this will take the sources in the src/ directory, compile them and link them into
+# lib(subdirectory_name).a in the build directory. This behaviour is entirely configurable,
+# please read the SDK documents if you need to do this.
+#
 
-COMPONENT_EMBED_TXTFILES := onboarding_config.json device_info.json update_root.pem update_public_key.pem
+COMPONENT_ADD_INCLUDEDIRS := .
