+++ main/device_control.h
@@ -1,42 +1,47 @@
-/* ***************************************************************************
+/**
+ *  @file    device_control.h
+ *  @author  Sean Mathews <coder@f34r.com>
+ *  @date    02/20/2020
  *
- * Copyright 2019 Samsung Electronics All Rights Reserved.
+ *  @brief Hardware abstraction
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  @copyright Copyright (C) 2020 Nu Tech Software Solutions, Inc.
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific
- * language governing permissions and limitations under the License.
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
- ****************************************************************************/
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
 
-
+// Defaults to ESP32 DEVKIT V4
+//#define CONFIG_TARGET_OLIMEX_ESP32_EVB
 //#define CONFIG_TARGET_WEMOS_D1_R32
 #ifdef CONFIG_TARGET_WEMOS_D1_R32
-
-#define GPIO_INPUT_BUTTON 18
-
-#define GPIO_OUTPUT_MAINLED 16
-#define GPIO_OUTPUT_MAINLED_0 26 /* use as ground */
-
-#define GPIO_OUTPUT_NOUSE1 17
-#define GPIO_OUTPUT_NOUSE2 25
+ #define GPIO_INPUT_BUTTON 18
+ #define GPIO_OUTPUT_MAINLED 16
+ #define GPIO_OUTPUT_MAINLED_0 26 /* use as ground */
+ #define GPIO_OUTPUT_NOUSE1 17
+ #define GPIO_OUTPUT_NOUSE2 25
+#elif CONFIG_TARGET_OLIMEX_ESP32_EVB
+ #define GPIO_INPUT_BUTTON 0
+ #define GPIO_OUTPUT_MAINLED 12
+ #define GPIO_OUTPUT_MAINLED_0 26 /* use as ground */
+ #define GPIO_OUTPUT_NOUSE1 14
+ #define GPIO_OUTPUT_NOUSE2 27
 #else // ESP32_DEVKITC_V4
-
-#define GPIO_INPUT_BUTTON 0
-
-#define GPIO_OUTPUT_MAINLED 12
-#define GPIO_OUTPUT_MAINLED_0 26 /* use as ground */
-
-#define GPIO_OUTPUT_NOUSE1 14
-#define GPIO_OUTPUT_NOUSE2 27
-
+ #define GPIO_INPUT_BUTTON 0
+ #define GPIO_OUTPUT_MAINLED 12
+ #define GPIO_OUTPUT_MAINLED_0 26 /* use as ground */
+ #define GPIO_OUTPUT_NOUSE1 14
+ #define GPIO_OUTPUT_NOUSE2 27
 #endif
 
 enum switch_onoff_state {
